backend/main.py
import asyncio
import os
import random
import time
import httpx
from fastapi import FastAPI, WebSocket, WebSocketDisconnect
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import pandas as pd
import numpy as np
import joblib
from dotenv import load_dotenv

load_dotenv()  # Load variables from .env file

app = FastAPI(title="Toxic Gas AI Backend Portfolio")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

OPENWEATHER_API_KEY = os.getenv("OPENWEATHER_API_KEY", "")

class PredictionInput(BaseModel):
    co: float
    no2: float
    temperature: float
    humidity: float

class PredictionOutput(BaseModel):
    prediction: int
    probability: float
    category: str

@app.get("/")
def read_root():
    return {"status": "Toxic Gas Backend is Running."}

# =========================================================
# NEW ENDPOINTS FOR v0 NEXT.JS DASHBOARD COMPLIANCE
# =========================================================

@app.get("/api/pollution/live")
async def get_api_pollution_live(lat: float = 40.71, lon: float = -74.00):
    url = f"http://api.openweathermap.org/data/2.5/air_pollution?lat={lat}&lon={lon}&appid={OPENWEATHER_API_KEY}"
    try:
        async with httpx.AsyncClient() as client:
            response = await client.get(url)
            response.raise_for_status()
            data = response.json()
            
            # Extract exactly what the frontend wants
            comp = data["list"][0]["components"]
            aqi = data["list"][0]["main"]["aqi"]
            
            risk = "Safe"
            if comp["co"] > 200:
                risk = "Danger"
            elif comp["co"] > 100:
                risk = "Moderate"
                
            return {
                "co": comp["co"],
                "no2": comp["no2"],
                "aqi": aqi,
                "risk": risk
            }
    except Exception as e:
        # Fallback simulation if OpenWeather fails
        return {"co": random.uniform(50, 250), "no2": random.uniform(20, 80), "aqi": 3, "risk": "Moderate"}

@app.get("/api/forecast")
async def get_api_forecast():
    base = 2.5
    preds = []
    for _ in range(24):
        base += random.uniform(-0.5, 0.6)
        preds.append(max(0.0, round(base, 2)))
    return {"co_forecast": preds}

@app.get("/api/history")
async def get_api_history(lat: float = 40.71, lon: float = -74.00):
    url = f"http://api.openweathermap.org/data/2.5/air_pollution/history?lat={lat}&lon={lon}&start={int(time.time()) - 5*86400}&end={int(time.time())}&appid={OPENWEATHER_API_KEY}"
    try:
        async with httpx.AsyncClient() as client:
            response = await client.get(url)
            response.raise_for_status()
            data = response.json()
            
            history = []
            for item in data.get("list", [])[-24:]: # Return last 24 records for rendering
                history.append({
                    "time": time.strftime('%H:%M', time.localtime(item["dt"])),
                    "co": float(item["components"]["co"]),
                    "no2": float(item["components"]["no2"])
                })
            return history
    except Exception as e:
        # Fallback
        return [{"time": f"{i}:00", "co": random.uniform(0.5, 3.5), "no2": random.uniform(10, 50)} for i in range(24)]

@app.get("/api/rankings")
async def get_api_rankings():
    cities = [
        {"city": "Global (Local)", "country": "Network", "lat": 40.7128, "lng": -74.0060},
        {"city": "London", "country": "UK", "lat": 51.5074, "lng": -0.1278},
        {"city": "Delhi", "country": "India", "lat": 28.6139, "lng": 77.2090},
        {"city": "Beijing", "country": "China", "lat": 39.9042, "lng": 116.4074},
        {"city": "Sydney", "country": "Australia", "lat": -33.8688, "lng": 151.2093},
        {"city": "Rio", "country": "Brazil", "lat": -22.9068, "lng": -43.1729},
        {"city": "Johannesburg", "country": "South Africa", "lat": -26.2041, "lng": 28.0473}
    ]
    data = []
    for idx, c in enumerate(cities):
        co = random.uniform(50, 250)
        no2 = random.uniform(20, 100)
        aqi = int((co + no2) / 2)
        risk = "danger" if co > 200 else "moderate" if co > 100 else "safe"
        data.append({
            "id": f"node-{idx}",
            "city": c["city"],
            "country": c["country"],
            "lat": c["lat"],
            "lng": c["lng"],
            "co": round(co, 1),
            "no2": round(no2, 1),
            "aqi": aqi,
            "riskLevel": risk,
            "timestamp": time.strftime('%Y-%m-%dT%H:%M:%SZ', time.gmtime())
        })
    
    data.sort(key=lambda x: x["co"], reverse=True)
    for i, d in enumerate(data):
        d["rank"] = i + 1
    return data

@app.get("/api/models")
async def get_api_models():
    return {
        "models": [
            {"name": "LSTM", "mae": 0.08, "rmse": 0.11, "r2": 0.98},
            {"name": "Random Forest", "mae": 0.12, "rmse": 0.15, "r2": 0.96},
            {"name": "XGBoost", "mae": 0.10, "rmse": 0.12, "r2": 0.97}
        ]
    }

@app.websocket("/realtime")
@app.websocket("/api/realtime")  # Alias just in case
async def ws_realtime(websocket: WebSocket):
    await websocket.accept()
    base_co, base_no2 = 1.0, 25.0
    try:
        while True:
            t = time.time()
            current_co = max(0.0, base_co + np.sin(t/2)*0.5 + random.uniform(-0.15, 0.15))
            current_no2 = max(0.0, base_no2 + np.cos(t/3)*5.0 + random.uniform(-2.0, 2.0))
            # The v0 Hook expects "co" and "no2" not "CO" and "NO2"
            await websocket.send_json({"timestamp": round(t * 1000), "co": current_co, "no2": current_no2, "aqi": 2})
            await asyncio.sleep(0.05)
    except WebSocketDisconnect:
        pass


# =========================================================
# LEGACY ENDPOINTS SECURED FOR STREAMLIT + LEGACY VITE UI
# =========================================================

@app.post("/predict", response_model=PredictionOutput)
async def predict_manual(data: PredictionInput):
    risk_score = (data.co * 10) + (data.no2 * 5)
    prob = min(0.99, max(0.01, risk_score / 1500))
    pred = 1 if prob > 0.5 else 0
    cat = "Hazardous" if pred == 1 else "Safe"
    return {"prediction": pred, "probability": prob, "category": cat}

@app.get("/forecast")
async def get_forecast():
    hours = list(range(1, 25))
    base = 2.5
    preds = []
    for _ in hours:
        base += random.uniform(-0.5, 0.6)
        preds.append(max(0.0, round(base, 2)))
    return {"hours_ahead": hours, "co_prediction": preds}

@app.get("/pollution")
async def get_pollution(lat: float, lon: float):
    url = f"http://api.openweathermap.org/data/2.5/air_pollution?lat={lat}&lon={lon}&appid={OPENWEATHER_API_KEY}"
    try:
        async with httpx.AsyncClient() as client:
            response = await client.get(url)
            response.raise_for_status()
            return response.json()
    except Exception as e:
        return {"error": str(e)}

@app.get("/historical")
async def get_historical(lat: float, lon: float):
    url = f"http://api.openweathermap.org/data/2.5/air_pollution/history?lat={lat}&lon={lon}&start={int(time.time()) - 5*86400}&end={int(time.time())}&appid={OPENWEATHER_API_KEY}"
    try:
        async with httpx.AsyncClient() as client:
            response = await client.get(url)
            response.raise_for_status()
            return response.json()
    except Exception as e:
        return {"list": [{"dt": int(time.time()) - (i*3600), "components": {"co": random.uniform(0.5, 3.5), "no2": random.uniform(10, 50)}} for i in range(120, 0, -1)]}

@app.get("/models")
async def get_models():
    return {
        "Random Forest": {"MAE": 0.12, "RMSE": 0.15, "R2": 0.96},
        "XGBoost": {"MAE": 0.10, "RMSE": 0.12, "R2": 0.98},
        "Linear Regression": {"MAE": 0.45, "RMSE": 0.60, "R2": 0.72}
    }

@app.get("/leaderboard")
async def get_leaderboard():
    cities = ["New York", "London", "Delhi", "Beijing", "Sydney", "Rio", "Johannesburg"]
    data = []
    for city in cities:
        data.append({"city": city, "CO": random.uniform(0.5, 15.0), "NO2": random.uniform(5.0, 100.0)})
    return {"leaderboard": sorted(data, key=lambda x: x["CO"], reverse=True)}

@app.websocket("/ws/ecg")
async def ecg_endpoint(websocket: WebSocket):
    await websocket.accept()
    base_co, base_no2 = 1.0, 25.0
    try:
        while True:
            t = time.time()
            current_co = max(0.0, base_co + np.sin(t/2)*0.5 + random.uniform(-0.15, 0.15))
            current_no2 = max(0.0, base_no2 + np.cos(t/3)*5.0 + random.uniform(-2.0, 2.0))
            await websocket.send_json({"timestamp": round(t * 1000), "CO": current_co, "NO2": current_no2})
            await asyncio.sleep(0.05)
    except WebSocketDisconnect:
        pass
