backend/requirements.txt
fastapi
uvicorn
websockets
pandas
numpy
scikit-learn
joblib
xgboost
python-dotenv
httpx
pydantic
